refactor: :truck: Move file to components/controls for reuse across all webviews
