+++ src/webview/view/components/controls/index.ts
@@ -1,1 +1,2 @@
 export * from './MultiSelect';
+export * from './LabelWithTooltip';
